Adiona ValorViewModel e Atualiza Models e Controllers

+++ src/CaixaEletronico.Api/Models/Conta.cs
@@ -3,5 +3,5 @@
 public class Conta
 {
     public int Id { get; set; }
-    public float Saldo { get; set; }
+    public double Saldo { get; set; }
 }
